Control0908

+++ modules/control/control_pid.h
@@ -22,26 +22,26 @@ static float ki_speed = 0;
 static float kd_speed = 0;
 // static float delta_u_torque;
 // static float u_pre_torque = 0;
-static float u_torque;
+static double u_torque;
 
 // static float steer_motor_spd;
 // static float tire_steer_spd;
 // static float steer_angle=0;
 static float veh_spd = 0;
 
-static float speed_motor_deadzone = 0;
+static double speed_motor_deadzone = 0;
 
-static float pid_integral = 0;
-static float pid_error = 0;
-static float pid_error_pre = 0;
+static double pid_integral = 0;
+static double pid_error = 0;
+static double pid_error_pre = 0;
 // static float fmottq = 0;
 // static float canbus_veh_spd = 0;
 // static int veh_mode = 0;
 
 float pid_speed(float veh_spd, float desire_v, float spd_motor_deadzone) {
   cout << "pid中的期望速度" << desire_v << endl;
-  // veh_spd = chassis_->speed_mps();
-  pid_error = desire_v - veh_spd;  // pid输入为当前车速误差
+  // pid输入为当前车速误差
+  pid_error = desire_v - veh_spd;
   cout << "纵向速度偏差" << pid_error << endl;
 
   pid_integral += pid_error;
